requirements.txt
huggingface-hub
paddleocr
paddlepaddle
db-sqlite3
flask
opencv-python
ultralytics
mysql-connector-python
flask-login
werkzeug